hw12-webserver Fixes
